Refactor CanMoveHere method to use float parameters for width and height

+++ src/main/java/utils/HelpMethods.java
@@ -20,7 +20,7 @@ public class HelpMethods {
         return false;
     }
 
-    public static boolean CanMoveHere(float x, float y, int width, int height, int[][] lvlData) {
+    public static boolean CanMoveHere(float x, float y, float width, float height, int[][] lvlData) {
         if (!IsSolid(x, y, lvlData))
             if (!IsSolid(x + width, y + height, lvlData))
                 if (!IsSolid(x + width, y, lvlData))
